Move inital users creation to factories
